+++ about.py
@@ -18,7 +18,7 @@ class BPYPlus:
     """
 
     _current_version = (1, 6, 0)
-    _blender_version = ((2, 91), (2, 92), (2, 93), (3, 0), (3, 1), (3, 2))
+    _blender_version = ((2, 93), (3, 0), (3, 1), (3, 2))
 
     @classmethod
     def version(cls) -> tuple: